refactor(card-creator): relocate export utilities to `utils/export.ts` and update imports
