+++ src/apps/card-creator/components/gallery/CardThumbnail.tsx
@@ -6,7 +6,7 @@ import {
 	type StoredCard,
 } from "@fabkit/apps/card-creator/persistence/card-storage";
 import { useCardCreator } from "@fabkit/apps/card-creator/stores/card-creator";
-import { rotateBlob } from "@fabkit/shared/export.ts";
+import { rotateBlob } from "@fabkit/apps/card-creator/utils/export.ts";
 import { useNavigate } from "@tanstack/react-router";
 import { Download, Edit, RotateCcw, Trash2 } from "lucide-react";
 import { useEffect, useMemo, useState } from "react";
